+++ CHANGES.md
@@ -7,7 +7,7 @@
 * `MoreCollectors.first/head` collectors are short-circuiting now.
 * `MoreCollectors.groupingByEnum` collector may short-circuit if downstream collector is short-circuiting.
 * `MoreCollectors.pairing` collector may short-circuit if both downstream collectors are short-circuiting.
-* Added new short-circuiting collectors: `onlyOne`, `intersecting`, `andingInt`, `andingLong`.
+* Added new short-circuiting collectors: `onlyOne`, `intersecting`, `toEnumSet`, `andingInt`, `andingLong`.
 * Added new collectors: `filtering`, `groupingBy` (with domain specification) which short-circuit when downstream collector is short-circuiting.
 * Added collectors `mapping`, `collectingAndThen`, `partitioningBy` which mimic standard JDK collectors, but short-circuit when downstream collector is short-circuiting.
 * Added `indexOf` methods for all stream types.